requirements.txt
streamlit
pillow
requests
beautifulsoup4
gTTS
moviepy==1.0.3